+++ make/photon/db/registry.sql
@@ -254,6 +254,41 @@ create table properties (
  UNIQUE (k)
  );
 
+create table harbor_label (
+ id int NOT NULL AUTO_INCREMENT,
+ name varchar(128) NOT NULL,
+ description text,
+ color varchar(16),
+# 's' for system level labels
+# 'u' for user level labels
+ level char(1) NOT NULL,
+# 'g' for global labels
+# 'p' for project labels
+ scope char(1) NOT NULL,
+ project_id int,
+ creation_time timestamp default CURRENT_TIMESTAMP,
+ update_time timestamp default CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP,
+ PRIMARY KEY(id),
+ CONSTRAINT unique_name_and_scope UNIQUE (name,scope)
+ );
+
+create table harbor_resource_label (
+ id int NOT NULL AUTO_INCREMENT,
+ label_id int NOT NULL,
+# the resource_id is the ID of project when the resource_type is p
+# the resource_id is the ID of repository when the resource_type is r
+# the resource_id is the name of image when the resource_type is i
+ resource_id varchar(256) NOT NULL,
+# 'p' for project
+# 'r' for repository
+# 'i' for image
+ resource_type char(1) NOT NULL,
+ creation_time timestamp default CURRENT_TIMESTAMP,
+ update_time timestamp default CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP,
+ PRIMARY KEY(id),
+ CONSTRAINT unique_label_resource UNIQUE (label_id,resource_id, resource_type)
+ );
+
 CREATE TABLE IF NOT EXISTS `alembic_version` (
     `version_num` varchar(32) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;