+++ src/common/models/base.go
@@ -32,5 +32,7 @@ func init() {
 		new(ClairVulnTimestamp),
 		new(WatchItem),
 		new(ProjectMetadata),
-		new(ConfigEntry))
+		new(ConfigEntry),
+		new(Label),
+		new(ResourceLabel))
 }